追加した8役職の処理を追加

+++ src/main/kotlin/com/ort/firewolf/application/service/SkillService.kt
@@ -13,7 +13,7 @@ class SkillService {
             CDef.Skill.listAll().filterNot {
                 CDef.Skill.listOfSomeoneSkill().contains(it)
             }.sortedBy {
-                it.order()
+                it.order().toInt()
             }.map { Skill(it) }
         )
     }
